pyproject.toml
[project]
name = "baby-ai-cam"
version = "0.1.0"
description = "AI-powered baby monitoring system with vision analysis and Discord alerts"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.3",
    "flask-socketio>=5.4.1",
    "opencv-python>=4.10.0",
    "numpy>=1.26.4",
    "librosa>=0.10.2",
    "discord-webhook>=1.3.1",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.5.0",
]
